created new contrib package
All extensions, including the extension loaders, are moved to this
new package.

This prevents attempts to load non-extension modules.
