+++ rackspace_cinder_extensions/__init__.py
@@ -1,38 +1,12 @@
-#  Copyright 2013-2016 Rackspace US, Inc.
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#
-#  http://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
-
-from oslo_config import cfg
 from oslo_log import log as logging
 
-from cinder.api import extensions
-
-# import registers global options
-from rackspace_cinder_extensions.common import config  # noqa
+from rackspace_cinder_extensions.api import contrib
 
 
-CONF = cfg.CONF
 LOG = logging.getLogger(__name__)
 
 
-def standard_extensions(ext_mgr):
-    extensions.load_standard_extensions(ext_mgr, LOG, __path__, __package__)
-
-
-def select_extensions(ext_mgr):
-    extensions.load_standard_extensions(ext_mgr, LOG, __path__, __package__,
-                                        CONF.rsapi_volume_ext_list)
-
-
 def rax_extensions(ext_mgr):
-    standard_extensions(ext_mgr)
+    LOG.warning('The rackspace_cinder_extensions.rax_extensions loader '
+                'is deprecated and will be removed in the v0.9 release.')
+    contrib.standard_extensions(ext_mgr)
