+++ rackspace_cinder_extensions/api/contrib/snapshot_progress.py
